+++ pkg/lm/ngram_model_test.go
@@ -149,7 +149,7 @@ func TestBinaryMarshalling(t *testing.T) {
 
 	// Create a decoder and receive a value.
 	input, _ := outDir.OpenInput("lm")
-	actual := &nGramModel{}
+	actual := NewNGramModel()
 
 	if _, err = actual.Load(input); err != nil {
 		t.Errorf("Unexpected error: %v", err)
